Updates to 0-1 plotting to accommodate changes to simulation output format
No longer need to find max and mean values from output - these are
calculated in the FORTRAN program.

+++ scripts/zeroonemultiplot.py
@@ -23,7 +23,7 @@ else:
 for i in datafolders:
     zerooneplotting.zeroonemeasurements(i)
 
-if os.path.exists(os.path.join(datafolders[0],"0-1data/radius0-1.txt")):
+if os.path.exists(os.path.join(datafolders[0],"0-1data/epiradius0-1.txt")):
     combined_radius = np.genfromtxt(os.path.join(datafolders[0],"0-1data/radius0-1.txt"))
     for i in range(1,len(datafolders)):
         data = np.genfromtxt(os.path.join(datafolders[i],"0-1data/radius0-1.txt"))
@@ -46,9 +46,43 @@ if os.path.exists(os.path.join(datafolders[0],"0-1data/radius0-1.txt")):
     ax1 = fig1.add_subplot(111)
     ax1.errorbar(outbin_edges,outmean,yerr=outstd,color="Red",ls="none")
     ax1.scatter(outbin_edges,outmean,color="Red")
-    ax1.set_title("Mean distance of primitive endoderm from centre of mass as a\nproportion of maximum possible value, averaged over "+str(len(datafolders))+" runs",y=1.05)
+    #ax1.set_title("Mean distance of epiblast from centre of mass as a\nproportion minimum possible value, averaged over "+str(len(datafolders))+" runs",y=1.05)
+    ax1.set_title("State of system as a proportion of the difference between perfectly sorted and random,\nas measured by mean distance of epiblast from centre of mass,\naveraged over "+str(len(datafolders))+" runs",y=1.05)
     ax1.set_xlabel("Time")
-    ax1.set_ylabel("Distance from centre of mass as a proportion of maximum")
+    #ax1.set_ylabel("Distance from centre of mass as a proportion of maximum")
+    ax1.set_ylabel("Sorting Index")
+    ax1.set_ylim([-0.1,1.1])
+    fig1.savefig(os.path.join(plotfolder,"radius.png"),bbox_inches="tight")
+    np.savetxt(os.path.join(datafolder,"radius.txt"),np.stack((outbin_edges,outmean,outstd),axis=1))
+
+if os.path.exists(os.path.join(datafolders[0],"0-1data/preradius0-1.txt")):
+    combined_radius = np.genfromtxt(os.path.join(datafolders[0],"0-1data/radius0-1.txt"))
+    for i in range(1,len(datafolders)):
+        data = np.genfromtxt(os.path.join(datafolders[i],"0-1data/radius0-1.txt"))
+        combined_radius = np.vstack((combined_radius,data))
+
+    zerodata = np.array([])
+    for i in range(0,len(combined_radius)):
+        if combined_radius[i,0] == 0.0:
+            zerodata = np.append(zerodata,combined_radius[i,1])
+
+    radius_mean,bin_edges,binnumber = binned_statistic(combined_radius[:,0],combined_radius[:,1],bins=5)
+    radius_zeromean = np.mean(zerodata)
+    radius_std,bin_edges,binnumber = binned_statistic(combined_radius[:,0],combined_radius[:,1],statistic=sem,bins=5)
+    radius_zerostd = sem(zerodata)
+
+    outbin_edges = np.append([0],bin_edges[0:5]+(bin_edges[1]-bin_edges[0])/2.0)
+    outmean = np.append(radius_zeromean,radius_mean)
+    outstd = np.append(radius_zerostd,radius_std)
+    fig1 = plt.figure()
+    ax1 = fig1.add_subplot(111)
+    ax1.errorbar(outbin_edges,outmean,yerr=outstd,color="Red",ls="none")
+    ax1.scatter(outbin_edges,outmean,color="Red")
+    #ax1.set_title("Mean distance of primitive endoderm from centre of mass as a\nproportion of maximum possible value, averaged over "+str(len(datafolders))+" runs",y=1.05)
+    ax1.set_title("State of system as a proportion of the difference between perfectly sorted and random,\nas measured by mean distance of primitive endoderm from centre of mass,\naveraged over "+str(len(datafolders))+" runs",y=1.05)
+    ax1.set_xlabel("Time")
+    #ax1.set_ylabel("Distance from centre of mass as a proportion of maximum")
+    ax1.set_ylabel("Sorting Index")
     ax1.set_ylim([-0.1,1.1])
     fig1.savefig(os.path.join(plotfolder,"radius.png"),bbox_inches="tight")
     np.savetxt(os.path.join(datafolder,"radius.txt"),np.stack((outbin_edges,outmean,outstd),axis=1))
@@ -76,9 +110,11 @@ if os.path.exists(os.path.join(datafolders[0],"0-1data/surface0-1.txt")):
     ax1 = fig1.add_subplot(111)
     ax1.errorbar(outbin_edges,outmean,yerr=outstd,color="Red",ls="none")
     ax1.scatter(outbin_edges,outmean,color="Red")
-    ax1.set_title("Proportion of external surface occupied by primitive endoderm as a\nproportion of maximum possible value, averaged over "+str(len(datafolders))+" runs",y=1.05)
+    #ax1.set_title("Proportion of external surface occupied by primitive endoderm as a\nproportion of maximum possible value, averaged over "+str(len(datafolders))+" runs",y=1.05)
+    ax1.set_title("State of system as a proportion of the difference between perfectly sorted and random,\nas measured by proportion of external surface occupied by primitive endoderm,\naveraged over "+str(len(datafolders))+" runs",y=1.05)
     ax1.set_xlabel("Time")
-    ax1.set_ylabel("Proportion of external surface occupied by primitive\nendoderm as a proportion of maximum possible value")
+    #ax1.set_ylabel("Proportion of external surface occupied by primitive\nendoderm as a proportion of maximum possible value")
+    ax1.set_ylabel("Sorting Index")
     ax1.set_ylim([-0.1,1.1])
     fig1.savefig(os.path.join(plotfolder,"surface.png"),bbox_inches="tight")
     np.savetxt(os.path.join(datafolder,"surface.txt"),np.stack((outbin_edges,outmean,outstd),axis=1))
@@ -106,9 +142,11 @@ if os.path.exists(os.path.join(datafolders[0],"0-1data/neighbour0-1.txt")):
     ax1 = fig1.add_subplot(111)
     ax1.errorbar(outbin_edges,outmean,yerr=outstd,color="Green",ls="none")
     ax1.scatter(outbin_edges,outmean,color="Green")
-    ax1.set_title("Number of epiblast-epiblast neighbour pairs as a proportion of maximum\npossible value, averaged over "+str(len(datafolders))+" runs",y=1.05)
+    #ax1.set_title("Number of epiblast-epiblast neighbour pairs as a proportion of maximum\npossible value, averaged over "+str(len(datafolders))+" runs",y=1.05)
+    ax1.set_title("State of system as a proportion of the difference between perfectly sorted and random,\nas measured by number of epiblast-epiblast neighbour pairs,\naveraged over "+str(len(datafolders))+" runs",y=1.05)
     ax1.set_xlabel("Time")
-    ax1.set_ylabel("Number of epiblast-epiblast pairs\nas a proportion of maximum")
+    #ax1.set_ylabel("Number of epiblast-epiblast pairs\nas a proportion of maximum")
+    ax1.set_ylabel("Sorting Index")
     ax1.set_ylim([-0.1,1.1])
     fig1.savefig(os.path.join(plotfolder,"neighbour.png"),bbox_inches="tight")
     np.savetxt(os.path.join(datafolder,"neighbour.txt"),np.stack((outbin_edges,outmean,outstd),axis=1))
